Add Netezza support
Extend DatabaseConnector.R connect.default with another if statement for
`dbms == "netezza"` that loads the netezza driver and makes a
connection. Amend the R documentation for the connect and
createConnectionDetails functions with Netezza-specific information.
Unfortunately, it looks like the driver is under licensing, so users will
have to drop their own driver into the `inst/java` directory. As such,
this commit adds relevant lines to the README.md and README.html files
as well as the R documentation in the DatabaseConnector.R file.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>