Github issue #75: Check SQL names for reserved words before inserting tables in to a database.
This commit adds a couple new folders to the package. The dev (development) folder contains a csv file with SQL reservered word and code to add that data to data folder so it is available when the package is loaded. The commit contains a new function to check SQL names.